requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
cryptography
python-dotenv